+++ src/components/Blog/Blog.scss
@@ -1,5 +1,5 @@
-@use "../../../assets/scss/variables.scss" as *;
-@use "../../../assets/scss/mixins.scss" as *;
+@use "../../assets/scss/variables.scss" as *;
+@use "../../assets/scss/mixins.scss" as *;
 
 .blog {
   padding: 1.5rem;
