+++ src/components/Home/ArticleForm/ArticleForm.scss
@@ -18,6 +18,7 @@
     display: flex;
     justify-content: center;
     align-items: center;
+
     height: 70%;
   }
