+++ src/components/Home/Home.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import "./Home.scss";
 
-import Header from "../common/Header/Header";
 import HomeBlog from "./HomeBlog/HomeBlog";
 import ArticleForm from "./ArticleForm/ArticleForm";
