+++ src/components/common/Article/Article.scss
@@ -31,10 +31,11 @@
     top: 0;
     left: 50%;
 
-    transform: translateX(-50%);
-    object-fit: cover;
     width: 100%;
     height: 100%;
+
+    transform: translateX(-50%);
+    object-fit: cover;
     z-index: 0;
   }
