+++ src/components/common/Footer/Footer.scss
@@ -13,7 +13,9 @@
   &__copyright {
     position: relative;
     top: 0.1875rem;
+
     font-size: 1.25rem;
+
     margin-right: 0.625rem;
   }
 }
