+++ src/components/common/Header/Header.scss
@@ -13,11 +13,13 @@
     left: 0;
     right: 0;
     bottom: 0;
+
     z-index: -1;
   }
 
   &__bg-filter {
     height: 100%;
+
     background-color: rgba(0, 0, 0, 0.4);
   }
 
@@ -73,20 +75,19 @@
     position: relative;
 
     &--active::before {
-      content: "";
       position: absolute;
       left: 0;
       bottom: -0.0625rem;
-      z-index: -1;
 
       width: 100%;
       height: 0.125rem;
 
       background-color: $color-accent;
+      z-index: -1;
+      content: "";
     }
 
     &:not(.header__link--active)::before {
-      content: "";
       position: absolute;
       right: 0;
       bottom: -0.125rem;
@@ -95,6 +96,7 @@
       height: 0.125rem;
 
       background-color: $color-accent;
+      content: "";
       transition: width 100ms ease-in;
     }
 
@@ -118,6 +120,7 @@
   }
   @include vp-tablet {
     min-height: 11.25rem;
+
     padding-bottom: 1.25rem;
 
     &__heading {
@@ -159,8 +162,8 @@
       background-color: transparent;
       border: 0;
       cursor: pointer;
-
       z-index: 5;
+
       @include transition-normal;
 
       @include interaction-effect-normal;
@@ -196,11 +199,11 @@
           }
 
           &::before {
+            position: absolute;
+            left: 0.125rem;
+
             width: 1.25rem;
             height: 0.1875rem;
-            position: absolute;
-
-            left: 0.125rem;
 
             background-color: $color-background-primary;
             content: "";
@@ -215,12 +218,12 @@
 
         &__nav {
           display: block;
+
           position: fixed;
           left: 0;
           top: 0;
           right: 0;
           bottom: 0;
-          z-index: 2;
 
           width: 100vw;
           height: 100vh;
@@ -228,6 +231,7 @@
           padding: 10rem 2rem 0;
 
           background-color: $color-background-tertiary;
+          z-index: 2;
         }
 
         &__nav-list {
