+++ src/components/common/Spinner/Spinner.scss
@@ -1,28 +1,28 @@
 @use "../../../assets/scss/variables.scss" as *;
 
-.lds-ring {
+.spinner {
   display: inline-block;
   position: relative;
 }
-.lds-ring div {
+.spinner div {
   box-sizing: border-box;
   display: block;
   position: absolute;
   border: 10px solid $color-accent;
   border-radius: 50%;
-  animation: lds-ring 1.2s cubic-bezier(0.5, 0, 0.5, 1) infinite;
+  animation: spinner 1.2s cubic-bezier(0.5, 0, 0.5, 1) infinite;
   border-color: $color-accent transparent transparent transparent;
 }
-.lds-ring div:nth-child(1) {
+.spinner div:nth-child(1) {
   animation-delay: -0.45s;
 }
-.lds-ring div:nth-child(2) {
+.spinner div:nth-child(2) {
   animation-delay: -0.3s;
 }
-.lds-ring div:nth-child(3) {
+.spinner div:nth-child(3) {
   animation-delay: -0.15s;
 }
-@keyframes lds-ring {
+@keyframes spinner {
   0% {
     transform: rotate(0deg);
   }
