+++ src/index.tsx
@@ -3,9 +3,6 @@ import ReactDOM from "react-dom/client";
 import "./assets/scss/style.scss";
 import { BrowserRouter as Router } from "react-router-dom";
 import App from "./App";
-// import { Provider } from "react-redux";
-// import { store } from "./app/store";
-import reportWebVitals from "./reportWebVitals";
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
@@ -13,15 +10,8 @@ const root = ReactDOM.createRoot(
 
 root.render(
   <React.StrictMode>
-    {/* <Provider store={store}> */}
     <Router>
       <App />
     </Router>
-    {/* </Provider> */}
   </React.StrictMode>
 );
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
