+++ src/utils/token.ts
@@ -1,4 +1,3 @@
 const TOKEN = "pj11daaQRz7zUIH56B9Z";
-// const TOKEN = "22";
 
 export default TOKEN;
